==> production/01-preparacion.Rmd <==
---
title: "Preparación de datos"
date: "`r format(Sys.time(), '%A %d %B %Y %X')`"
output:
  html_document:
    toc: yes
    toc_float:
      collapsed: yes
      smooth_scroll: no
      number_sections: yes
    code_folding: hide
    number_sections: yes
editor_options:
  chunk_output_type: console
---

# Setup

```{r setup}
knitr::opts_chunk$set(
	echo = TRUE,
	message = FALSE,
	warning = FALSE
)
options(scipen=9999) # desactivar notacion cientifica
remove(list = ls()) #limpieza del entorno de trabajo
```

```{css, echo=FALSE}
# /* expandir el espacio del documento*/
body .main-container {
      max-width: 1600px !important;
      width: 1600px !important;
    }
#/* expandir width de la TOC*/
div.tocify {
    width: 22% !important;
    max-width: 331px !important;
    max-height: 102% !important;
}
```

## Individual data

```{r}
remove(list = ls()) # clean workspace
# .rs.restartR() #restart R session
if (!require("pacman")) install.packages("pacman") # install pacman
# load libraries
pacman::p_load(dplyr,readxl,sjmisc, sjlabelled, questionr, car,here)
load(here::here("input/data-original/elsoc_long_2016_2023.Rdata"))

elsoc_long_2016_2023[elsoc_long_2016_2023 ==-999] <- NA
elsoc_long_2016_2023[elsoc_long_2016_2023 ==-888] <- NA
elsoc_long_2016_2023[elsoc_long_2016_2023 ==-777] <- NA
elsoc_long_2016_2023[elsoc_long_2016_2023 ==-666] <- NA

# __________________________________________________
elsoc_long_2016_2023 <- 
elsoc_long_2016_2023 %>% 
  #create new variables
  mutate(part_marchas=c08_02,
         part_rrss=c08_04,
         disp_marchas=c09_02,
         disp_rrss=c09_04,
         mov_soc=c20,
         part_electoral1=c11, # presidenciales 2013 y 2017
         part_electoral3=c43, # Plebiscito nueva const
         part_electoral4=c50, # presidenciales primera vuelta
         part_electoral5=c52, # presidenciales segunda vuelta
         interes_pol=c13,
         voto_influye=c10_02,
         autoritarismo1=c18_04,
         autoritarismo2=c18_05,
         autoritarismo3=c18_06,
         autoritarismo4=c18_07,
         autoritarismo5=c18_08,
         adop_homo=c37_01,
         aborto=c37_02,
         restriccion_migrantes=c37_05,
         pensiones=c37_04,
         perc_des=c18_11,
         rol_estado_educ=c37_03,
         sexo=m0_sexo,
         edad=m0_edad,
         educacion=m01,
         ola
         ) %>% 
  sjlabelled::drop_labels(., drop.na = FALSE)
  
#
sjmisc::frq(elsoc_long_2016_2023$marchar)
elsoc_long_2016_2023$marchar <- 
sjlabelled::set_label(x = elsoc_long_2016_2023$marchar, 
                      label = "Frecuencia: marchar")

# Socioeconomic_________________________________________________________________

# Education_______________________________________
 elsoc_long_2016_2023$educ <- 
   car::recode(elsoc_long_2016_2023$educacion,
               "c(1,2,3,4,5)=1;c(6,7)=2;c(8,9,10)=3; c(-888,-999)=NA")
 elsoc_long_2016_2023$educ <-
   factor(elsoc_long_2016_2023$educ,
          labels = c("Media o menos","Técnica","Universitaria"))

#reverse education, reference level is the highest level
#elsoc_long_2016_2023$educ <- forcats::fct_rev(elsoc_long_2016_2023$educ)

 elsoc_long_2016_2023$educ <- 
 sjlabelled::set_label(x = elsoc_long_2016_2023$educ,
                       label = "Educación")
 sjmisc::frq(elsoc_long_2016_2023$educ)
 
 
elsoc_long_2016_2023$universitaria <- 
   car::recode(elsoc_long_2016_2023$educacion,
               "c(1,2,3,4,5,6,7)=1;c(8,9,10)=2; c(-888,-999)=NA")
 elsoc_long_2016_2023$universitaria <-
   factor(elsoc_long_2016_2023$universitaria,
          labels = c("Menos que universitaria","Universitaria"))

#reverse education, reference level is the highest level
#elsoc_long_2016_2023$educ <- forcats::fct_rev(elsoc_long_2016_2023$educ)

 sjmisc::frq(elsoc_long_2016_2023$universitaria)
 

# Household income_________________________________________

#Impute midpoint of income ranges
# elsoc_long_2016_2023$m30_rec <-
# as.numeric(car::recode(elsoc_long_2016_2023$m30,
#            "1=110000;2=251000;3=305000;4=355000;5=400000;
#             6=445000;7=490000;8=535000;9=585000;10=640000;11=700000;12=765000;
#             13=845000;14=935000;15=1040000;16=1180000;17=1375000;18=1670000;
#             19=2275000;20=2700000;NA=NA;c(-888,-999)=NA"))
# 
# #Impute midpoint of income ranges (2021)
# elsoc_long_2016_2023$m30b_rec <-
# as.numeric(car::recode(elsoc_long_2016_2023$m30b,
#            "1=125000;2=300000;3=400000;4=575000;5=70000;NA=NA;c(-888,-999)=NA"))
# 
# sjmisc::frq(elsoc_long_2016_2023$m30b_rec)
# 
# #Recode DK/DA of Income to NA
# elsoc_long_2016_2023$m29_rec <-
#   as.numeric(car::recode(elsoc_long_2016_2023$m29,"c(-888,-999)=NA"))
# 
# #replace NA of income with new imputed variable
# elsoc_long_2016_2023$m29_imp <- 
#   ifelse(test = !is.na(elsoc_long_2016_2023$m29_rec),
#          yes =  elsoc_long_2016_2023$m29_rec,
#          no =  elsoc_long_2016_2023$m30_rec)
# summary(elsoc_long_2016_2023$m29_imp)
# 
# elsoc_long_2016_2023$m29_imp <- 
#   ifelse(test = is.na(elsoc_long_2016_2023$m29_imp),
#          yes =  elsoc_long_2016_2023$m30b_rec,
#          no =  elsoc_long_2016_2023$m29_imp)
# summary(elsoc_long_2016_2023$m29_imp)
# 
# # deflate at each year's prices
# elsoc_long_2016_2023$deflactor <-
#   with(elsoc_long_2016_2023, case_when(
#     ola == 2016 ~ 113.88 / 123.82,
#     ola == 2017 ~ 116.46 / 123.82,
#     ola == 2018 ~ 119.45 / 123.82,
#     ola == 2019 ~ 123.82 / 123.82
#   ))
# 
# # N Household:
# elsoc_long_2016_2023 <-
#   elsoc_long_2016_2023 %>%
#   mutate(n_hogar =
#            dplyr::case_when(ola == 1 ~ nhogar1,
#                             ola == 2 ~ m46_nhogar,
#                             ola == 3 ~ m54,
#                             ola == 4 ~ m54,
#                             ola == 5 ~ m54,
#                             ola == 6 ~ m54,
#                             ola == 7 ~ m54))
# sjmisc::frq(elsoc_long_2016_2023$n_hogar)
# 
# #Recode DK/DA to NA
# elsoc_long_2016_2023$n_hogar_r<-
#   car::recode(elsoc_long_2016_2023$n_hogar,"c(-888,-999)=NA")
# 
# # Per capita household income:
# elsoc_long_2016_2023$ing_pc <- 
#   (elsoc_long_2016_2023$m29_imp/elsoc_long_2016_2023$n_hogar_r)
# 
# elsoc_long_2016_2023$ing_pc <-
# sjlabelled::set_label(x = elsoc_long_2016_2023$ing_pc,
#                       label = "Ingreso per capita del hogar")  
# 
# sjmisc::descr(elsoc_long_2016_2023$ing_pc)
# 
# # Compute income quintiles
# elsoc_long_2016_2023 <- elsoc_long_2016_2023 %>% 
#   group_by(ola) %>% 
#   mutate(quintil = ntile(-desc(ing_pc), 5)) %>% 
#   ungroup()
# 
# elsoc_long_2016_2023$quintil <- 
#   factor(elsoc_long_2016_2023$quintil,
#          levels = c(1, 2, 3, 4, 5),
#          labels = c('Q1', 'Q2', 'Q3', 'Q4', 'Q5')) # Quintiles as factors
# 
# #reverse quintile, reference level is the highest quintile
# #elsoc_long_2016_2023$quintil <- forcats::fct_rev(elsoc_long_2016_2023$quintil)
# 
# elsoc_long_2016_2023$quintil <- 
# sjlabelled::set_label(x = elsoc_long_2016_2023$quintil,
#                       label = "Household income quintile per capita")  
# 
# sjmisc::frq(elsoc_long_2016_2023$quintil)
# 
# #include new quintile category with missing cases
# elsoc_long_2016_2023$quintil1<-
#   car::recode(elsoc_long_2016_2023$quintil, 
#               "'Q1'='Q1';'Q2'= 'Q2';'Q3'='Q3';'Q4'='Q4';'Q5'='Q5'; NA='QNA'")
# 
# #elsoc_long_2016_2023$quintil1 <- factor(elsoc_long_2016_2023$quintil1, c("Q1","Q2","Q3","Q4","Q5","QNA"))
# 
# elsoc_long_2016_2023$quintil1 <- 
# sjlabelled::set_label(x = elsoc_long_2016_2023$quintil1,
#                       label = "Household income quintile per capita (NA)") 
# sjmisc::frq(elsoc_long_2016_2023$quintil1)


#Age Ranges
elsoc_long_2016_2023$edad <- as.numeric(elsoc_long_2016_2023$m0_edad)

elsoc_long_2016_2023$edad_cat <- 
  factor(car::recode(elsoc_long_2016_2023$m0_edad, 
                     "18:29=1;30:49=2;50:64=3;65:150=4"),
         labels = c('18-29', '30-49', '50-64', '65 or more'))
elsoc_long_2016_2023$edad_cat <-
  sjlabelled::set_label(elsoc_long_2016_2023$edad_cat, 
                        label = c("Rango etario")) 

# Political position ____________________________
elsoc_long_2016_2023$pos_id <-
factor(
  car::recode(
    elsoc_long_2016_2023$c15,
    "c(11,12,-888,-999)='No se identifica';c(0,1,2,3,4)='Izquierda';
     c(5)='Centro';c(6,7,8,9,10)='Derecha'"
  ),
  levels = c('Izquierda', 'Centro', 'Derecha', 'No se identifica')
)

elsoc_long_2016_2023$pos_id <- factor(elsoc_long_2016_2023$pos_id,
                                      levels = levels(elsoc_long_2016_2023$pos_id))

elsoc_long_2016_2023$pos_id <-
sjlabelled::set_label(x = elsoc_long_2016_2023$pos_id,
                      label = "Identificación política")

sjmisc::frq(elsoc_long_2016_2023$pos_id)

```


```{r}
# Select variables______________________________________________________________

df_study1 <- 
elsoc_long_2016_2023 %>% 
  filter(muestra == 1) %>% #keep original sample (vs refresh)
  select(idencuesta,ola,
         muestra,
         ponderador_long_total, segmento, estrato,
         sexo,
         educ,
         universitaria,
         # cuestion_mig,
         part_marchas,
         part_rrss,
         disp_marchas,
         disp_rrss,
         mov_soc,
         part_electoral1,
         part_electoral3,
         part_electoral4,
         part_electoral5,
         interes_pol,
         voto_influye,
         autoritarismo1,
         autoritarismo2,
         autoritarismo3,
         autoritarismo4,
         autoritarismo5,
         adop_homo,
         aborto,
         restriccion_migrantes,
         pensiones,
         perc_des,
         rol_estado_educ,
         #educ,educyear,
         # ing_pc,
         # quintil,
         #quintil1,
         # ess_inm,
         # nation_t1,
         # know_inm,
         # know_inm_bi,
         # frien_inm_bi_t4,
         # frien_inm_bi_t1,
         # know_inm_t1t4,
         # frien_inm,
         # frien_inm_bi,
         # frien_inm_t1t4,
         pos_id,
         edad,
         m37_01, m37_02,
         m48)


# Reshape long to wide
df_study1_wide <- df_study1 %>% 
  tidyr::pivot_wider(id_cols = c("idencuesta","muestra"),
                     names_from = "ola",
                     # names_prefix = c("educ","educyear","ing_pc","quintil","quintil1"),
                     values_from = names(select(df_study1,ponderador_long_total,segmento, estrato,sexo,educ,universitaria, part_marchas:m48)))
```


```{r}
# fix data to w01 values

df_study1_wide$edad_2 <-df_study1_wide$edad_1 #age
df_study1_wide$edad_3 <-df_study1_wide$edad_1
df_study1_wide$edad_4 <-df_study1_wide$edad_1
df_study1_wide$edad_5 <-df_study1_wide$edad_1
df_study1_wide$edad_6 <-df_study1_wide$edad_1
df_study1_wide$edad_7 <-df_study1_wide$edad_1

df_study1_wide$sexo_2 <-df_study1_wide$sexo_1 #sex
df_study1_wide$sexo_3 <-df_study1_wide$sexo_1
df_study1_wide$sexo_4 <-df_study1_wide$sexo_1
df_study1_wide$sexo_5 <-df_study1_wide$sexo_1
df_study1_wide$sexo_6 <-df_study1_wide$sexo_1
df_study1_wide$sexo_7 <-df_study1_wide$sexo_1
# 
df_study1_wide$educ_2 <-df_study1_wide$educ_1 #education
df_study1_wide$educ_3 <-df_study1_wide$educ_1
df_study1_wide$educ_4 <-df_study1_wide$educ_1
df_study1_wide$educ_5 <-df_study1_wide$educ_1
df_study1_wide$educ_6 <-df_study1_wide$educ_1
df_study1_wide$educ_7 <-df_study1_wide$educ_1
# 
df_study1_wide$universitaria_2 <-df_study1_wide$universitaria_1 #univeristaria
df_study1_wide$universitaria_3 <-df_study1_wide$universitaria_1
df_study1_wide$universitaria_4 <-df_study1_wide$universitaria_1
df_study1_wide$universitaria_5 <-df_study1_wide$universitaria_1
df_study1_wide$universitaria_6 <-df_study1_wide$universitaria_1
df_study1_wide$universitaria_7 <-df_study1_wide$universitaria_1


# df_study1_wide$quintil1_2 <-df_study1_wide$quintil1_1 #quintiles
# df_study1_wide$quintil1_3 <-df_study1_wide$quintil1_1
# df_study1_wide$quintil1_4 <-df_study1_wide$quintil1_1
# df_study1_wide$quintil1_5 <-df_study1_wide$quintil1_1
# df_study1_wide$quintil1_6 <-df_study1_wide$quintil1_1
# df_study1_wide$quintil1_7 <-df_study1_wide$quintil1_1

df_study1_wide$pos_id_2 <-df_study1_wide$pos_id_1 # political position
df_study1_wide$pos_id_3 <-df_study1_wide$pos_id_1
df_study1_wide$pos_id_4 <-df_study1_wide$pos_id_1
df_study1_wide$pos_id_5 <-df_study1_wide$pos_id_1
df_study1_wide$pos_id_6 <-df_study1_wide$pos_id_1
df_study1_wide$pos_id_7 <-df_study1_wide$pos_id_1

# Las variables de disposición solo están en ola 1, 3 y 6

df_study1_wide$part_electoral_1 <- df_study1_wide$part_electoral1_1
df_study1_wide$part_electoral_2 <- df_study1_wide$part_electoral1_1
df_study1_wide$part_electoral_3 <- df_study1_wide$part_electoral1_3
df_study1_wide$part_electoral_4 <- df_study1_wide$part_electoral1_3
df_study1_wide$part_electoral_5 <- df_study1_wide$part_electoral3_5
df_study1_wide$part_electoral_6 <- df_study1_wide$part_electoral4_6
df_study1_wide$part_electoral_7 <- df_study1_wide$part_electoral5_6

df_study1_wide$autoritarismo1_6 <- df_study1_wide$autoritarismo1_5
df_study1_wide$autoritarismo3_6 <- df_study1_wide$autoritarismo3_5
df_study1_wide$autoritarismo4_6 <- df_study1_wide$autoritarismo4_5

df_study1_wide$adop_homo_5 <- df_study1_wide$adop_homo_6

df_study1_wide$rol_estado_educ_5 <- df_study1_wide$rol_estado_educ_6

df_study1_wide <- df_study1_wide %>%
  mutate(
    hijo = case_when(
      m48_2 == 1 | m48_4 == 1 ~ 1,
      m37_01_1 > 0 | m37_02_1 > 0 | m37_01_3 > 0 | m37_02_3 > 0 ~ 1,
      is.na(m48_2) & is.na(m48_4) & 
      is.na(m37_01_1) & is.na(m37_02_1) & 
      is.na(m37_01_3) & is.na(m37_02_3) ~ NA_real_,
      TRUE ~ 0
    )
  ) %>%
  mutate(
    hijo_1 = hijo,
    hijo_2 = hijo,
    hijo_3 = hijo,
    hijo_4 = hijo,
    hijo_5 = hijo,
    hijo_6 = hijo,
    hijo_7 = hijo
  )

names(df_study1_wide)
dim(df_study1_wide)

sjPlot::view_df(df_study1_wide,
                 show.frq = T,show.values = T,show.na = T,show.prc = T, show.type = T)
```


```{r}
# reshape from long to wide
pacman::p_load(datasets,data.table)
df_study1_long <- data.table::melt.data.table(data.table::setDT(df_study1_wide),
              id.vars = c("idencuesta","muestra"),
              variable.name = c("ola"),
              measure = patterns("^ponderador_long_total_", "^segmento_", "^estrato_",
                                 "^part_marchas_", "^part_rrss_",
                                 "^mov_soc_",  "^part_electoral_",
                                 "^part_electoral1_", "^part_electoral3_", "^part_electoral4_", "^part_electoral5_",
                                 "^interes_pol_", "^voto_influye_", 
                                 "^autoritarismo1_", "^autoritarismo2_", "^autoritarismo3_", "^autoritarismo4_", "^autoritarismo5_", 
                                 "^adop_homo_", "^aborto_", "^restriccion_migrantes_", "^pensiones_",
                                 "^perc_des_", "^rol_estado_educ_",
                                 "^educ_", "^universitaria_", 
                     #"^quintil1_","^ess_",
                                 "^pos_id_",
                                 "^edad_", "^hijo_",
                     "^sexo_"
                                 ),
              value.name = c("ponderador_long_total", "segmento", "estrato",
                                 "part_marchas", "part_rrss",
                                 "mov_soc", "part_electoral",
                                 "part_electoral1", "part_electoral3", "part_electoral4", "part_electoral5",
                                 "interes_pol", "voto_influye",
                                 "autoritarismo1", "autoritarismo2", "autoritarismo3", "autoritarismo4", "autoritarismo5", 
                                 "adop_homo", "aborto", "restriccion_migrantes", "pensiones",
                                 "perc_des", "rol_estado_educ",
                                 "educ", "universitaria", 
                     #"^quintil1_","^ess_",
                                 "pos_id",
                                 "edad", "hijo",
                     "sexo"
                     ),
              na.rm = F,value.factor = TRUE
              )

df_study1_long$mov_estudiantil <- as.numeric(df_study1_long$mov_soc == 1)
df_study1_long$mov_feminista <- as.numeric(df_study1_long$mov_soc == 8)
names(df_study1_long) #check names of long dataset
dim(df_study1_long) #check dimensions of the dataframe
# Original dataset with 6 waves
df_study2_long <- df_study1_long

# # filter the dataset for the waves 1 to 4 and 6
# df_study1_long <-
# df_study1_long %>% 
#   filter(ola %in% c(1,2,3,4,6)) %>% 
#   mutate(ola=factor(ola,levels = 1:6,labels = 1:6))
# dim(df_study1_long) #check, now is OK

# df_study1_long <- 
# set_label(x = df_study1_long,
#           label = get_label(select(df_study1,names(df_study1_long))))
  

sjPlot::view_df(df_study1_long,
                show.frq = T,show.values = T,show.na = T,show.prc = T,
                show.type = T)
#______________________________________________________________________________
# obtain the idencuesta for wave 7
ids <- 
  df_study1 %>% 
  select(idencuesta,ola) %>% 
  filter(ola==7) %>% 
  sjmisc::frq(idencuesta,show.na = F) %>% as.data.frame()


# filter data by the idencuesta of t5
df_study1_long_t7 <- 
  df_study1_long %>%
  filter(idencuesta %in% ids$val)

names(df_study1_long_t7)
dim(df_study1_long_t7)
sjmisc::frq(df_study1_long_t7$ola)

# SAVE DATA____________________________________________________________________
save(df_study1_long,file = here::here("input/data-proc/df_study1_long.RData"))
save(df_study1_long_t7,file = here::here("input/data-proc/df_study1_long_t7.RData"))
save(df_study2_long,file = here::here("input/data-proc/df_study2_long.RData"))
# save codebook
sjPlot::view_df(df_study1_long,
        show.frq = T,
        show.prc = T,
        show.na = T,
        file = here::here("output/df_study1_long.html"))

sjPlot::view_df(df_study1_long_t7,
        show.frq = T,
        show.prc = T,
        show.na = T,
        file = here::here("output/df_study1_long_t7.html"))
```
